pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-knowledge-ingestion"
version = "0.1.0"
description = "Ingestion and indexing layer for an LLM data engineering platform"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "LLM Data Platform Team" }]
keywords = ["llm", "ingestion", "indexing", "data-engineering", "retrieval"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
llm-knowledge-ingest = "llm_knowledge_ingestion.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
show_error_codes = true
pretty = true
mypy_path = "src"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --strict-markers --disable-warnings --cov=llm_knowledge_ingestion --cov-report=term-missing"
testpaths = ["tests"]
